requirements.txt
streamlit
scikit-learn
pandas
numpy
pdfplumber
python-docx
pytesseract
Pillow
matplotlib
xgboost